// src/components/types.ts
export type ColorOptions =
  | 'purple700'
  | 'purple900'
  | 'lighBlue500'
  | 'pink400'
  | 'pink500'
  | 'red100'
  | 'red500'
  | 'green100'
  | 'green500'
  | 'green700'
  | 'yellow100'
  | 'yellow500'
  | 'white'
  | 'white100'
  | 'grey100'
  | 'grey500'
  | 'black600'
  | 'black700'
  | 'black900';